QCamera2: Fix compile error caused by ARRAY_SIZE() comparison

+++ QCamera2/HAL/QCamera2HWI.cpp
@@ -1395,7 +1395,7 @@ int QCamera2HardwareInterface::initCapabilities(int cameraId,mm_camera_vtbl_t *c
 {
     int rc = NO_ERROR;
     QCameraHeapMemory *capabilityHeap = NULL;
-    int i, x;
+    size_t i, x;
 
     /* Allocate memory for capability buffer */
     capabilityHeap = new QCameraHeapMemory(QCAMERA_ION_USE_CACHE);
